requirements.txt
woocommerce>=3.0.0
customtkinter>=5.2.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0 